# CLOUD-COMPUTING/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Mentalys REST API
  version: 0.0.1
  description: REST API for Mentalys Application developed by Abdi Setiawan - c179b4ky0013
  contact:
    name: Abdi Setiawan
    email: c179b4ky0013@bangkit.academy
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /psychiatrists:
    get:
      tags:
        - Psychiatrists
      summary: Mendapatkan daftar psikiater
      description: Mengambil semua psikiater yang tersedia.
      responses:
        '200':
          description: Daftar psikiater berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "5GvRbLGEeAsMNhLwOHv3"
                        fullName:
                          type: string
                          example: "Dr. Tifa Lockhart"
                        title:
                          type: string
                          example: "MD, Psychiatry"
                        mainRole:
                          type: string
                          example: "Psychiatric Consultant"
                        consultationFee:
                          type: number
                          example: 150000
        '500':
          description: Terjadi kesalahan saat membaca data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error reading data"
  /psychiatrists/{id}:
    get:
      tags:
        - Psychiatrists Details
      summary: Mendapatkan detail psikiater
      description: Mengambil informasi detail psikiater berdasarkan ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID psikiater yang ingin diambil
          schema:
            type: string
      responses:
        '200':
          description: Detail psikiater berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "5GvRbLGEeAsMNhLwOHv3"
                      fullName:
                        type: string
                        example: "Dr. Tifa Lockhart"
                      title:
                        type: string
                        example: "MD, Psychiatry"
                      aboutMe:
                        type: string
                        example: "Dr. Tifa is a compassionate psychiatrist with over a decade of experience in mental health care."
                      consultationFee:
                        type: number
                        example: 150000
        '404':
          description: Psikiater tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Psychiatrist not found"
        '500':
          description: Terjadi kesalahan saat membaca data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error reading data"
...